support selecting start and end point for shortest path, improve path styling
