+++ src/index.html
@@ -15,7 +15,7 @@
                 <div class="cell">
                     <div class="walls"></div>
                     <div class="vertex"></div>
-                    <div class="label"></div>
+                    <div class="path"></div>
                 </div>
             </div>
         </div>
